+++ core/runner/modifiers.go
@@ -22,7 +22,7 @@ const (
 // ModifyWithLock bulk modifies contacts by loading and locking them, applying modifiers and processing the resultant events.
 //
 // Note we don't load the user object from org assets as it's possible that the user isn't part of the org, e.g. customer support.
-func ModifyWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, userID models.UserID, contactIDs []models.ContactID, modifiersByContact map[models.ContactID][]flows.Modifier, includeTickets map[models.ContactID][]*models.Ticket) (map[*flows.Contact][]flows.Event, []models.ContactID, error) {
+func ModifyWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, userID models.UserID, contactIDs []models.ContactID, modifiersByContact map[models.ContactID][]flows.Modifier, includeTickets map[models.ContactID][]*models.Ticket, via models.Via) (map[*flows.Contact][]flows.Event, []models.ContactID, error) {
 	eventsByContact := make(map[*flows.Contact][]flows.Event, len(modifiersByContact))
 	remaining := contactIDs
 	start := time.Now()
@@ -32,7 +32,7 @@ func ModifyWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAsse
 			return nil, nil, ctx.Err()
 		}
 
-		es, skipped, err := tryToModifyWithLock(ctx, rt, oa, userID, remaining, modifiersByContact, includeTickets)
+		es, skipped, err := tryToModifyWithLock(ctx, rt, oa, userID, remaining, modifiersByContact, includeTickets, via)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -44,7 +44,7 @@ func ModifyWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAsse
 	return eventsByContact, remaining, nil
 }
 
-func tryToModifyWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, userID models.UserID, ids []models.ContactID, modifiersByContact map[models.ContactID][]flows.Modifier, includeTickets map[models.ContactID][]*models.Ticket) (map[*flows.Contact][]flows.Event, []models.ContactID, error) {
+func tryToModifyWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, userID models.UserID, ids []models.ContactID, modifiersByContact map[models.ContactID][]flows.Modifier, includeTickets map[models.ContactID][]*models.Ticket, via models.Via) (map[*flows.Contact][]flows.Event, []models.ContactID, error) {
 	// try to get locks for these contacts, waiting for up to a second for each contact
 	locks, skipped, err := clocks.TryToLock(ctx, rt, oa, ids, time.Second)
 	if err != nil {
@@ -70,7 +70,7 @@ func tryToModifyWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.Or
 		eventsByContact[scene.Contact] = make([]flows.Event, 0) // TODO only needed to avoid nulls until jsonv2
 
 		for _, mod := range modifiersByContact[scene.ContactID()] {
-			evts, err := scene.ApplyModifier(ctx, rt, oa, mod, userID)
+			evts, err := scene.ApplyModifier(ctx, rt, oa, mod, userID, via)
 			if err != nil {
 				return nil, nil, fmt.Errorf("error applying modifier: %w", err)
 			}
@@ -86,8 +86,8 @@ func tryToModifyWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.Or
 	return eventsByContact, skipped, nil
 }
 
-// BulkModify bulk modifies contacts without locking.. used during contact creation
-func BulkModify(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, userID models.UserID, mcs []*models.Contact, contacts []*flows.Contact, modifiers map[flows.ContactUUID][]flows.Modifier) (map[*flows.Contact][]flows.Event, error) {
+// BulkModify bulk modifies contacts without locking.. used during contact creation and imports
+func BulkModify(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, userID models.UserID, mcs []*models.Contact, contacts []*flows.Contact, modifiers map[flows.ContactUUID][]flows.Modifier, via models.Via) (map[*flows.Contact][]flows.Event, error) {
 	scenes := make([]*Scene, 0, len(mcs))
 	eventsByContact := make(map[*flows.Contact][]flows.Event, len(mcs))
 
@@ -97,7 +97,7 @@ func BulkModify(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, 
 		eventsByContact[contact] = make([]flows.Event, 0)
 
 		for _, mod := range modifiers[mc.UUID()] {
-			evts, err := scene.ApplyModifier(ctx, rt, oa, mod, userID)
+			evts, err := scene.ApplyModifier(ctx, rt, oa, mod, userID, via)
 			if err != nil {
 				return nil, fmt.Errorf("error applying modifier %T to contact %s: %w", mod, mc.UUID(), err)
 			}