+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+v26.1.6 (2026-01-15)
+-------------------------
+ * Never credit error events to users
+ * Don't persist URN-taken error events that occur via imports
+ * Record on events if they happened via import
+
+v26.1.5 (2026-01-15)
+-------------------------
+ * Persist URN taken error events
+ * Update to latest goflow which improves more error events
+
+v26.1.4 (2026-01-14)
+-------------------------
+ * Update to latest goflow which improves failure events
+ * Increase max steps per sprint to 250
+ * Persist failure events from failed sessions
+
 v26.1.3 (2026-01-13)
 -------------------------
  * Ignore created_on on event_received tasks and just use now